moved progess bar component from onboarding and demoExample to common folder
