+++ src/components/common/ProgressBar/ProgressSection/index.js
