+++ src/components/common/ProgressBar/ProgressSection/styles.js
