+++ src/components/common/ProgressBar/styles.js
